Add: added navigation bar in quran_screen refactor: in splash, tasbe7 and quran screens make the route name static and global to be sharable by other dart files.
